requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pgvector==0.2.0
python-dotenv==1.0.0
anthropic==0.5.0
sentence-transformers==2.2.2
pydantic==2.3.0
pytest==7.4.0
httpx==0.24.1
python-multipart==0.0.6 